py/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astroid==3.3.11
attrs==25.3.0
beautifulsoup4==4.13.4
boltons==21.0.0
bracex==2.6
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
colorama==0.4.6
cryptography==45.0.5
cssselect==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
face==24.0.0
fake-useragent==2.2.0
fastcore==1.8.5
glom==22.1.0
google==3.0.0
google-auth==2.40.3
google-genai==1.24.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
html-to-markdown==1.4.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
llms-txt==0.0.4
lxml==6.0.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
numpy==2.3.1
openai==1.97.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==25.0
pandas==2.3.1
parse==1.20.2
peewee==3.18.2
platformdirs==4.3.8
playwright==1.54.0
pluggy==1.6.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
PyNaCl==1.5.0
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-html==0.10.0
rich==13.5.3
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
semgrep==1.128.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
text-unidecode==1.3
tomli==2.0.2
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
wcmatch==8.5.2
websockets==10.4
wrapt==1.17.2
zipp==3.23.0
